updated code duplication
