+++ sources/packages/api/src/esdc-integration/e-cert-integration/e-cert-full-time-integration/e-cert-full-time-integration.service.ts
@@ -3,7 +3,6 @@ import { ConfigService, SshService } from "../../../services";
 import {
   combineDecimalPlaces,
   getDisbursementValuesByType,
-  getFieldOfStudyFromCIPCode,
   getGenderCode,
   getMaritalStatusCode,
   getTotalDisbursementAmount,
@@ -20,12 +19,12 @@ import { ECertFullTimeFileHeader } from "./e-cert-files/e-cert-file-header";
 import { ECertFullTimeFileFooter } from "./e-cert-files/e-cert-file-footer";
 import { ECertFullTimeFileRecord } from "./e-cert-files/e-cert-file-record";
 import {
+  DisbursementValue,
   DisbursementValueType,
   OfferingIntensity,
 } from "../../../database/entities";
 import { ECertIntegrationService } from "../e-cert-integration.service";
 import { ECertResponseRecord } from "../e-cert-files/e-cert-response-record";
-import { ECertDisbursementSchedule } from "../../../services/disbursement-schedule-service/disbursement-schedule.models";
 
 /**
  * Manages the file content generation and methods to
@@ -178,22 +177,8 @@ export class ECertFullTimeIntegrationService extends ECertIntegrationService {
     );
   }
 
-  /**
-   * Create the Full-Time e-Cert record with the information needed to generate the
-   * entire record to be sent to ESDC.
-   * @param disbursement disbursement that contains all information to
-   * generate the record.
-   * @returns e-Cert record.
-   */
-  createECertRecord(disbursement: ECertDisbursementSchedule): ECertRecord {
-    const now = new Date();
-    const application = disbursement.studentAssessment.application;
-    const addressInfo = application.student.contactInfo.address;
-    const offering = application.currentAssessment.offering;
-    const fieldOfStudy = getFieldOfStudyFromCIPCode(
-      offering.educationProgram.cipCode,
-    );
-    const awards = disbursement.disbursementValues
+  populateAwards(disbursementValues: DisbursementValue[]): Award[] {
+    return disbursementValues
       .filter(
         (disbursementValue) =>
           disbursementValue.valueType !== DisbursementValueType.BCGrant,
@@ -209,38 +194,5 @@ export class ECertFullTimeIntegrationService extends ECertIntegrationService {
                 : disbursementValue.valueAmount,
           } as Award),
       );
-
-    return {
-      sin: application.student.sin,
-      courseLoad: offering.courseLoad,
-      applicationNumber: application.applicationNumber,
-      documentNumber: disbursement.documentNumber,
-      disbursementDate: disbursement.disbursementDate,
-      documentProducedDate: now,
-      negotiatedExpiryDate: disbursement.negotiatedExpiryDate,
-      schoolAmount: disbursement.tuitionRemittanceRequestedAmount,
-      educationalStartDate: offering.studyStartDate,
-      educationalEndDate: offering.studyEndDate,
-      federalInstitutionCode: offering.institutionLocation.institutionCode,
-      weeksOfStudy: application.currentAssessment.assessmentData.weeks,
-      fieldOfStudy,
-      yearOfStudy: offering.yearOfStudy,
-      completionYears: offering.educationProgram.completionYears,
-      enrollmentConfirmationDate: disbursement.coeUpdatedAt,
-      dateOfBirth: application.student.birthDate,
-      lastName: application.student.user.lastName,
-      firstName: application.student.user.firstName,
-      addressLine1: addressInfo.addressLine1,
-      addressLine2: addressInfo.addressLine2,
-      city: addressInfo.city,
-      country: addressInfo.country,
-      provinceState: addressInfo.provinceState,
-      postalCode: addressInfo.postalCode,
-      email: application.student.user.email,
-      gender: application.student.gender,
-      maritalStatus: application.relationshipStatus,
-      studentNumber: application.studentNumber,
-      awards,
-    } as ECertRecord;
   }
 }
